unittest code refactor
